Extract TemplateRenderer from EMailSender
All rendering (placeholder values, paragraph split, auto-links, {logo},
highlighting, escaping) moves to the new Mail\TemplateRenderer with a
two-method API: renderSubject() and renderBody() returning [html,
plain] pairs ready for addBodyText(). EMailSender shrinks to
orchestration and sending; Defaults and IURLGenerator leave its
constructor.
